core: envelope: remove dead code / reorganize
