--- core/src/main/java/fr/sncf/osrd/envelope/EnvelopePartMeta.java
@@ -1,5 +1,0 @@
-package fr.sncf.osrd.envelope;
-
-/** A container for envelope metadata */
-public abstract class EnvelopePartMeta {
-}
