+++ core/src/main/java/fr/sncf/osrd/envelope/part/EnvelopePart.java
@@ -1,8 +1,11 @@
-package fr.sncf.osrd.envelope;
+package fr.sncf.osrd.envelope.part;
 
 import static fr.sncf.osrd.envelope.EnvelopePhysics.intersectStepWithSpeed;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import fr.sncf.osrd.envelope.EnvelopeAttr;
+import fr.sncf.osrd.envelope.EnvelopePhysics;
+import fr.sncf.osrd.envelope.SearchableEnvelope;
 import fr.sncf.osrd.utils.jacoco.ExcludeFromGeneratedCodeCoverage;
 import java.util.*;
 import java.util.stream.Collectors;
