+++ core/src/main/java/fr/sncf/osrd/envelope_sim/allowances/MarecoAllowance.java
@@ -5,9 +5,8 @@ import static fr.sncf.osrd.envelope_sim.overlays.EnvelopeDeceleration.decelerate
 import static java.lang.Math.*;
 
 import fr.sncf.osrd.envelope.*;
-import fr.sncf.osrd.envelope.constraint.ConstrainedEnvelopePartBuilder;
-import fr.sncf.osrd.envelope.constraint.EnvelopeCeiling;
-import fr.sncf.osrd.envelope.constraint.SpeedFloor;
+import fr.sncf.osrd.envelope.part.*;
+import fr.sncf.osrd.envelope.part.constraints.*;
 import fr.sncf.osrd.envelope_sim.EnvelopeProfile;
 import fr.sncf.osrd.envelope_sim.EnvelopeSimContext;
 import fr.sncf.osrd.envelope_sim.overlays.EnvelopeCoasting;
