+++ core/src/main/java/fr/sncf/osrd/envelope_sim/allowances/mareco_impl/CoastingOpportunity.java
@@ -1,10 +1,8 @@
 package fr.sncf.osrd.envelope_sim.allowances.mareco_impl;
 
 import fr.sncf.osrd.envelope.Envelope;
-import fr.sncf.osrd.envelope.EnvelopePart;
+import fr.sncf.osrd.envelope.part.EnvelopePart;
 import fr.sncf.osrd.envelope_sim.EnvelopeSimContext;
-import fr.sncf.osrd.envelope_sim.PhysicsPath;
-import fr.sncf.osrd.envelope_sim.PhysicsRollingStock;
 
 public interface CoastingOpportunity {
     /** Returns the location at which coasting shall end */
