+++ core/src/main/java/fr/sncf/osrd/train/StandaloneTrainSchedule.java
@@ -2,7 +2,6 @@ package fr.sncf.osrd.train;
 
 import com.carrotsearch.hppc.DoubleArrayList;
 import fr.sncf.osrd.envelope_sim.allowances.Allowance;
-
 import java.util.ArrayList;
 import java.util.Objects;
