+++ core/src/test/java/fr/sncf/osrd/envelope/EnvelopeShape.java
@@ -2,6 +2,7 @@ package fr.sncf.osrd.envelope;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import fr.sncf.osrd.envelope.part.EnvelopePart;
 import java.util.ArrayList;
 
 public enum EnvelopeShape {
