+++ core/src/test/java/fr/sncf/osrd/envelope_sim/MaxSpeedEnvelopeTest.java
@@ -4,6 +4,7 @@ import static fr.sncf.osrd.envelope.EnvelopeShape.*;
 import static fr.sncf.osrd.envelope_sim_infra.MRSP.LimitKind.SPEED_LIMIT;
 
 import fr.sncf.osrd.envelope.*;
+import fr.sncf.osrd.envelope.part.EnvelopePart;
 import fr.sncf.osrd.envelope_sim.pipelines.MaxSpeedEnvelope;
 import fr.sncf.osrd.envelope_sim_infra.MRSP.LimitKind;
 import fr.sncf.osrd.train.TestTrains;
